requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.26.156
botocore==1.29.156
build==0.10.0
CacheControl==0.12.14
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cryptography==41.0.1
debugpy==1.6.7
dependency-injector==4.41.0
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.5
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.97.0
fastapi-cache==0.1.0
filelock==3.12.2
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.3
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.17.3
jwt==1.3.1
keyring==23.13.1
kombu==5.3.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
mysqlclient==2.1.1
packaging==23.1
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.7.0
pluggy==1.0.0
poetry-core==1.6.1
poetry-plugin-export==1.4.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rapidfuzz==2.15.1
redis==4.5.4
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.6
webencodings==0.5.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
